organize docgen api in order to split tests (extract resolvers)
