+++ src/server/api/models/docgen/resolvers/index.js
@@ -1,0 +1,8 @@
+const componentResolver = require('./componentResolver')
+const indexResolver = require('./indexResolver')
+
+module.exports = Object.assign(
+  {},
+  { componentResolver },
+  { indexResolver }
+)
